Swe fixes (#359)
* Stops crashes if optional SWE time related elements aren't found

* Eliminate use of eval in SWE code, replace with module-level class dict

Eliminates an unprotected `eval` statement in the SWE code which could
lead to code execution from a specially crafted XML payload.  It was
possible to run code which followed the rules of an appropriately formed
XML tag.  This updates the code to instead introspect the module, and
maps the string names of the objects to their object counterpart for
safer dispatching.  Ideally, an explicit whitelist of classes should be
used, but it was uncertain which classes would possibly be called from
parsing the XML.

* Remove superfluous sys import

* Move sys import with other imports