Allow workspace qualified typenames in get_schema.
This commit allows the use of workspace qualified typenames (i.e.
'workspace:typename') in the get_schema method.

It is possible to have types with the same name in different workspaces,
but the returning schema only includes the non-qualified typename.

When passing a workspace qualified typename, we have to strip the
workspace when searching for the corresponding XML element.