Fix ISO metadata parsing for empty gmd:featureCatalogueCitation.
This commit fixes the ISO metadata parsing for records that contain an
empty gmd:featureCatalogueCitation element.

It also adds tests for ISO metadata parsing as to ensure not to break
existing functionality in the future. Tests are added for two ISO
metadata records: one from DOV (already existing in the testdata) and
one from GéoBretage.

Closes issue #491.