Refactor identification.resourcelanguage
Align resourcelanguage in MD_DataIdentification to language in
MD_Metadata.

In MD_Metadata we have 'language' for the language string and
'languagecode' for the language code.

In MD_DataIdentification we only had 'resourcelanguage' for the language
code. This commit refactors this to 'resourcelanguage' for the language
strings and 'resourcelanguagecode' for the language codes; analoguous to
the existing implementation in MD_Metadata.