Moved propertyname='*' from def to inside method
For WFS 1.1.0
With Post method now being available, having propertyname='*' as a
default value  causes problems for the post request. Default value
is now None, just like for 2.0.0.

Since I don't know if there are cases where this default was necessary,
added a condition when `method=get` to set propertynames to '*' is
it's not assigned a value.