Include custom headers in WMS requests
I recently needed to interact with a WMS service which require
authentication via a `User-agent` header (from France IGN [1]).

When using `WebMapService` you can actually provide custom headers but
they are not used later on when doing the actual HTTP requests:

    wms = WebMapService(
        'http://example.com/wms',
        version='1.3.0',
        headers={'User-agent': 'my-app/1.0'}
    )

    wms.getmap(...)  # Fails as not authorized

The `openURL()` helper already supports sending headers if provided, so
it was just a matter of adding the WMS client instance headers (if present) to
this function call.

Includes a small test.

[1] https://geoservices.ign.fr/documentation/donnees-ressources-wms-geoportail.html